# doc/api-docs-smart/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PinShop Server",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/group/initiate": {
      "post": {
        "summary": "发起拼团\n\n<p>\n用户成为团长，创建一个新的拼团会话。\n</p>",
        "deprecated": false,
        "description": "用户支付成功后调用此接口",
        "tags": [
          "GroupBuyController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "ruleId": {
                    "format": "int64",
                    "description": "拼团规则ID",
                    "type": "number",
                    "example": "0"
                  },
                  "userId": {
                    "format": "int64",
                    "description": "用户ID",
                    "type": "number",
                    "example": "0"
                  },
                  "skuId": {
                    "format": "int64",
                    "description": "商品SKU ID",
                    "type": "number",
                    "example": "0"
                  }
                },
                "required": [
                  "ruleId",
                  "userId",
                  "skuId"
                ]
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "&0&0&0"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/d4a93d5b69e349f7884016b0f02b403d"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": 0,\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-group-initiate-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/group/join": {
      "post": {
        "summary": "参与拼团\n\n<p>\n用户加入已有的拼团会话。\n</p>",
        "deprecated": false,
        "description": "用户支付成功后调用此接口",
        "tags": [
          "GroupBuyController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": {
                    "format": "int64",
                    "description": "拼团会话ID",
                    "type": "number",
                    "example": "0"
                  },
                  "userId": {
                    "format": "int64",
                    "description": "用户ID",
                    "type": "number",
                    "example": "0"
                  },
                  "skuId": {
                    "format": "int64",
                    "description": "商品SKU ID",
                    "type": "number",
                    "example": "0"
                  }
                },
                "required": [
                  "sessionId",
                  "userId",
                  "skuId"
                ]
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "&0&0&0"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/2e2e10f941a0ac19f25c4a7520d458b1"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-group-join-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/group/{skuId}/sessions": {
      "get": {
        "summary": "获取当前正在拼的团\n\n<p>\n返回指定商品正在进行的拼团会话列表。\n前端用于展示\"还有人在拼单\"功能。\n</p>",
        "deprecated": false,
        "description": "建议前端每10秒刷新一次，获取最新拼团状态",
        "tags": [
          "GroupBuyController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/8b51fe013d666dba3883617f4b677966"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": [\n    {\n      \"sessionId\": 0,\n      \"userAvatar\": \"\",\n      \"userName\": \"\",\n      \"missingNum\": 0,\n      \"timeLeft\": 0\n    }\n  ],\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-group-{skuId}-sessions-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "商品SKU ID",
            "required": true,
            "example": "0",
            "name": "skuId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/group/my": {
      "get": {
        "summary": "获取我的拼团记录\n\n<p>\n返回用户参与的所有拼团记录。\n</p>",
        "deprecated": false,
        "tags": [
          "GroupBuyController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/9410a0b91c4b61edc2c0cc3b326781a2"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"total\": 0,\n    \"groups\": [\n      {\n        \"sessionId\": 0,\n        \"skuId\": 0,\n        \"status\": 0,\n        \"statusText\": \"\",\n        \"currentCount\": 0,\n        \"targetCount\": 0,\n        \"timeLeft\": 0,\n        \"createTime\": \"yyyy-MM-dd HH:mm:ss\"\n      }\n    ]\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-group-my-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "用户ID",
            "required": true,
            "example": "0",
            "name": "userId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "页码（默认1）",
            "required": true,
            "example": "\"1\"",
            "name": "pageNum"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "每页大小（默认10）",
            "required": true,
            "example": "\"10\"",
            "name": "pageSize"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/group/{sessionId}/detail": {
      "get": {
        "summary": "获取拼团详情\n\n<p>\n返回指定拼团会话的详细信息，包括所有成员列表。\n</p>",
        "deprecated": false,
        "tags": [
          "GroupBuyController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/a19dfded3a48cf587ac88d99b2f0fb93"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"sessionId\": 0,\n    \"skuId\": 0,\n    \"status\": 0,\n    \"statusText\": \"\",\n    \"currentCount\": 0,\n    \"targetCount\": 0,\n    \"missingCount\": 0,\n    \"timeLeft\": 0,\n    \"expireTime\": \"yyyy-MM-dd HH:mm:ss\",\n    \"members\": [\n      {\n        \"userId\": 0,\n        \"userName\": \"\",\n        \"userAvatar\": \"\",\n        \"isInitiator\": true\n      }\n    ]\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-group-{sessionId}-detail-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "拼团会话ID",
            "required": true,
            "example": "0",
            "name": "sessionId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "8b51fe013d666dba3883617f4b677966": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/09c9eec20f91ed2cb5420140f0162ecf"
            }
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
      "2e2e10f941a0ac19f25c4a7520d458b1": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "type": "object"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "9410a0b91c4b61edc2c0cc3b326781a2": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/b89f91399864ff59bfc61ea087eb3599"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "a19dfded3a48cf587ac88d99b2f0fb93": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/bdfa4268c79ce5057c1654d968615dcf"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "d4a93d5b69e349f7884016b0f02b403d": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "format": "int64",
            "description": "业务数据",
            "type": "number"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "b89f91399864ff59bfc61ea087eb3599": {
        "properties": {
          "total": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "groups": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/70a35442deda44889df342102b732da2"
            }
          }
        }
      },
      "bdfa4268c79ce5057c1654d968615dcf": {
        "properties": {
          "sessionId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "skuId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "statusText": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "currentCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "targetCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "missingCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "timeLeft": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "expireTime": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "members": {
            "description": "No comments found.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/8dc98d17bc4be889f9b7d4a1279b8281"
            }
          }
        }
      },
...
      "8dc98d17bc4be889f9b7d4a1279b8281": {
        "properties": {
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "userName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "userAvatar": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "isInitiator": {
            "format": "boolean",
            "description": "No comments found.",
            "type": "boolean",
            "example": "true"
          }
        }
      },
...
      "70a35442deda44889df342102b732da2": {
        "properties": {
          "sessionId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "skuId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "statusText": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "currentCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "targetCount": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "timeLeft": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "createTime": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
      "09c9eec20f91ed2cb5420140f0162ecf": {
        "properties": {
          "sessionId": {
            "format": "int64",
            "description": "拼团会话ID",
            "type": "number",
            "example": "0"
          },
          "userAvatar": {
            "format": "string",
            "description": "团长头像URL",
            "type": "string"
          },
          "userName": {
            "format": "string",
            "description": "团长昵称（脱敏显示，如\"张***\"）",
            "type": "string"
          },
          "missingNum": {
            "format": "int32",
            "description": "还差几人成团",
            "type": "integer",
            "example": "0"
          },
          "timeLeft": {
            "format": "int64",
            "description": "剩余时间（毫秒）",
            "type": "number",
            "example": "0"
          }
        }
      },
...
      "da0c932b0e6d3b90859201037a7929ee": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "type": "object"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
    }
  }
}